+++ composer.json
@@ -11,7 +11,8 @@
     "require": {
         "php" : ">=7.2",
         "ext-pdo" : "*",
-        "twig/twig": "^2.4"
+        "twig/twig": "^2.4",
+        "symfony/webpack-encore-bundle": "^1.7"
     },
     "require-dev": {
         "phpmd/phpmd": "^2.7",
