+++ src/View/Admin/_eventAdmin.html.twig
@@ -19,11 +19,11 @@
                     <td>{{ id.title }}</td>
                     <td>{{ id.date }}</td>
                     <td>{{ id.location }}</td>
-                    <td><a class="mini activity-button btn btn-primary btn-lg" href="showEvent/{{ id.id }}">Modifier</a>
+                    <td><a class="mini activity-button btn btn-primary btn-lg" href="/./AdminEvent/showEvent/{{ id.id }}">Modifier</a>
                     </td>
                     <td>
                         <button>
-                            <a href="delete/{{ id.id }}"><img class="delete_button" src="/./assets/icons/logo_delete.png" alt="logo_delete"></a>
+                            <a href="/./AdminEvent/delete/{{ id.id }}"><img class="delete_button" src="/./assets/icons/logo_delete.png" alt="logo_delete"></a>
                         </button>
                     </td>
                 </tr>
@@ -32,6 +32,6 @@
                 </tbody>
             </table>
         </div>
-    <a class="activity-button btn btn-primary btn-lg" href="/./Admin/createEvent">Ajouter un évènement</a>
+    <a class="activity-button btn btn-primary btn-lg" href="/./AdminEvent/createEvent">Ajouter un évènement</a>
   </section>
