+++ src/View/Admin/_lessonAdmin.html.twig
@@ -28,7 +28,7 @@
                     </td>
                     <td>
                         <button>
-                            <a href="/./AdminLesson/delete/{{ id.id }}"><img class="delete_button" src="/./assets/icons/logo_delete.png" alt="logo_delete"></a>
+                            <a href="/./AdminLesson/delete/{{ lesson.id }}"><img class="delete_button" src="/./assets/icons/logo_delete.png" alt="logo_delete"></a>
                         </button>
                     </td>
                 </tr>
