+++ src/View/Admin/eventFormLayout.html.twig
@@ -3,7 +3,7 @@
 {% block content %}
     <h1>{% block title %}{% endblock %}</h1>
     <section class="form">
-        <form class = "event-form" method="post" action="/./Admin/addEvent">
+        <form class = "event-form" method="post" action="/./AdminEvent/addEvent">
             {% include 'Admin/_eventDetails.html.twig' %}
             <div class="form-button-pannel">
                 {% block buttons %}{% endblock %}
